Self Referral Form
This commit is all of our current updates for the 61-self-referral-form. Right now this page handles input validation and stores all fields in php variables to be used in a stored procedure. We are currently debugging and testing the stored procedure, but until it is finalized the SQL has been commented out for now.

More updates soon to come!